fix: the database configurations are now stored in their respective attributes
For some reason, I created an attribute called `configuration` which would store
configurations for all kinds of databases. This design choice moves away from the
way the data is stored in the database. Also, the client expects the configuration
values to be in their respective attributes. Therefore, all the relevant types
were updated.