feat: created utility function `getUserInfo`
The `getUserInfo` utility function plays a central role in Google OAuth2
authentication. It accepts an authorization code returned from Google.
A combination of the authorization code and client secret is used to
fetch the access code. At this point, the OAuth2 protocol is completed.
Finally, a request is sent to Google API (authorized by the access token)
to fetch the user info and returned to the caller.