fix: fixed weird behavior manifested by `import`
When `resolver` is invoked it returns the screen component. Inside `useEffect`,
the result of `resolver` works as expected, that is, a component type. However,
when it was stored to `target` state, the `target` state contained the rendered
component instead of the component type!

As a work around, two new states were created for storing the required information
from the component type. These states are directly updated within the `useEffect`.